Title: 정수 사각형 최대 합, Time: 11ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240821/정수 사각형 최대 합/maximum-sum-path-in-square.cpp
@@ -7,15 +7,15 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int n;
     cin >> n;
-    vector<vector<int>> v(n, vector<int>(n, 0));
-    for(int i = 0; i<n; i++)
+    vector<vector<long long>> v(n, vector<long long>(n, 0));
+    for(long long i = 0; i<n; i++)
     {
-        for(int j = 0; j<n; j++)
+        for(long long j = 0; j<n; j++)
         {
             cin >> v[i][j];
         }
     }
-    vector<vector<int>> dp(101, vector<int>(101, 0));
+    vector<vector<long long>> dp(101, vector<long long>(101, 0));
     dp[0][0] = v[0][0];
     for(int i = 1; i<n; i++)
     {
